change thredded overrides to make zeitwerk compatible
as mentioned in https://github.com/fxn/zeitwerk/issues/240#issuecomment-1328312345
