--- app/helpers/thredded/application_helper.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-require_dependency File.expand_path("../../app/helpers/thredded/application_helper", Thredded::Engine.called_from)
-module Thredded
-  module ApplicationHelper
-    #     include Thredded::Engine.routes.url_helpers
-    #     include Thredded::Workgroup::Engine.routes.url_helpers
-    #
-  end
-end
