--- app/helpers/thredded/workgroup/application_helper.rb
@@ -1,8 +1,0 @@
-# frozen_string_literal: true
-
-module Thredded
-  module Workgroup
-    module ApplicationHelper
-    end
-  end
-end
