+++ app/view_models/thredded/workgroup/topics_page_view_with_last_post.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Thredded
+  module Workgroup
+    module TopicsPageViewWithLastPost
+      protected
+
+      def refine_scope(topics_page_scope)
+        super.includes(:last_post)
+      end
+    end
+  end
+end
