+++ lib/overrides/thredded/all_view_hooks_override.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+Thredded::AllViewHooks.class_eval do
+  include Thredded::Workgroup::ViewHooks
+end
