+++ lib/overrides/thredded/posts_controller_override.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+Thredded::PostsController.class_eval do
+  prepend ::Thredded::Workgroup::PostsControllerWhichRedirects
+end
