+++ lib/overrides/thredded/topics_page_view_override.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+Thredded::TopicsPageView.class_eval do
+  prepend ::Thredded::Workgroup::TopicsPageViewWithLastPost
+end
