+++ lib/thredded/workgroup/version.rb
@@ -2,6 +2,6 @@
 
 module Thredded
   module Workgroup
-    VERSION = "0.3.1"
+    VERSION = "0.4.0"
   end
 end
